Make Hardcoded Unsigned Int Explicit in size() comparisons.
I've had numerous issues with segfaults and bugs showing up when not using explicitly unsigned integers in these comparisons when 'size()' is being called on a vector that has not been initialized properly. This happens often in the code when the vector in question is not intended to be used, but I still need to know I can skip it.

Making hardcoded unsigned integers explicit allows the comparisons to evaluate correctly.